refactor: move utils to their own file
