+++ apps/src/code-studio/pd/workshop_dashboard/WorkshopFormTemplate/index.tsx
@@ -1,6 +1,6 @@
 import Alert from '@code-dot-org/component-library/alert';
 import {Heading1} from '@code-dot-org/component-library/typography';
-import moment from 'moment-timezone';
+import {isEmpty} from 'lodash';
 import React, {
   FC,
   useCallback,
@@ -14,7 +14,6 @@ import {useParams} from 'react-router-dom';
 import {useFetch} from '@cdo/apps/util/useFetch';
 
 import {workshopLabel} from '../utils/workshopLabel';
-import {DATE_FORMAT, TIME_FORMAT} from '../workshopConstants';
 
 import {generateNewSession} from './components/SessionsEditor';
 import {sessionsReducer} from './reducers/sessionsReducer';
@@ -28,62 +27,16 @@ import PublishSettings from './sections/PublishSettings';
 import Schedule from './sections/Schedule';
 import {
   FieldConfig,
-  Session,
   SessionFormState,
   Workshop,
   WorkshopFormState,
   WorkshopFormTemplateProps,
 } from './types';
+import {workshopDataToState, sessionDataToState} from './utils';
 
 import styles from './styles.module.scss';
 
 const REQUIRED_ERROR = 'Required';
-
-export const isEmpty = (
-  value: string | boolean | number | unknown[] | null | undefined
-) =>
-  value === undefined ||
-  value === null ||
-  value === '' ||
-  (Array.isArray(value) && value.length === 0);
-
-export const workshopDataToState = (data: Workshop): WorkshopFormState => ({
-  course: data.course ?? '',
-  capacity: data.capacity?.toString() ?? '',
-  description: data.description ?? '',
-  facilitators: data.facilitators ?? [],
-  fee: data.fee ?? '',
-  grades: data.grades ?? [],
-  hidden: data.hidden ?? false,
-  name: data.name ?? '',
-  notes: data.notes ?? '',
-  organizerId: data.organizer?.id ?? null,
-  prereq: data.prereq ?? '',
-  hasPrereq: data.prereq ? true : false,
-  regionalPartnerId: data.regional_partner_id ?? null,
-  registrationLink: data.registration_link ?? '',
-  subject: data.subject ?? '',
-  suppressEmail: data.suppress_email ?? false,
-  courseOfferings: data.course_offerings?.map(n => n.toString()) ?? [],
-  participantGroupType: data.participant_group_type ?? '',
-  timeZone: data.time_zone ?? Intl.DateTimeFormat().resolvedOptions().timeZone,
-});
-
-export const sessionDataToState = (
-  data: Session[],
-  timeZone: string
-): SessionFormState[] =>
-  data.map(session => ({
-    id: `existing-${session.id}`,
-    date: moment(session.start).tz(timeZone).format(DATE_FORMAT),
-    start: moment(session.start).tz(timeZone).format(TIME_FORMAT),
-    end: moment(session.end).tz(timeZone).format(TIME_FORMAT),
-    locationAddress: session.location_address ?? '',
-    locationName: session.location_name ?? '',
-    meetingLink: session.meeting_link ?? '',
-    format: session.session_format ?? 'in_person',
-    sameAsPrevious: false,
-  }));
 
 export const WorkshopFormTemplate: FC<WorkshopFormTemplateProps> = ({
   config,
